refactor: JS 코드 분리
